process_avocado_filters: add script to generate vt-only-filters
For s390x, we want to use a minimal subset of the Cartesian product
of all variants in tp-libvirt that have been previously defined via
vt-only-filters and vt-no-filters.

Add a script that takes that input and creates a new vt-only-filter
applicable on s390-only and writes it to a file ending in "-s390x.only".

Furthermore, add the minor change to update avocado's prefix for vt
test cases with a note about the legacy value.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>